# Ajoute un index sur les événements pour optimiser le temps des requêtes