--- a/db/migrate/20201228154851_add_index_on_evenements_session_id.rb
+++ b/db/migrate/20201228154851_add_index_on_evenements_session_id.rb
@@ -1,0 +1,5 @@
+class AddIndexOnEvenementsSessionId < ActiveRecord::Migration[6.0]
+  def change
+    add_index :evenements, :session_id
+  end
+end