Co-authored-by: Marine Dominé <marine.domine@beta.gouv.fr>

Pour https://trello.com/c/zJm0yVYk/356-certaines-structures-constatent-un-blocage-sur-laccueil-pendant-les-%C3%A9valuations

Contexte:

On charge les événements pour chaque partie ( probablement pour les compétences fortes ). La table événements est la plus grosse, et la recherche prend ~ 150ms sur mon poste ( quasi 200ms sur le serveur ).

Cette PR ajoute un index qui accélère considérablement cette requête SQL.
Sur mon poste, cette modification a fait passé la requête de 2000ms à 250ms.

Je pense que ça peut valoir le coup d'aller plus loin mais ça fait déjà un beau gain pour une seule PR ;-)